+++ changelog.md
@@ -12,6 +12,7 @@
   - add chunk sorting option to the lazy collection `set_processing` method
 - Replace `gen_dataset/train.jsonl` with the original fictitious templates and the dataset generation script.
 - Update the README with the instructions to download the public pre-trained model.
+- Improve packaging to add evaluation results to the model's meta field and packaged model README (for HF)
 
 ## v0.3.0 - 2023-12-01